# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cellier API",
...
    "contact": {
      "name": "Equipo Cellier"
    },
    "license": {
      "name": "Uso interno"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Renovar el par de tokens",
        "description": "Canjea un refresh token vigente por un par nuevo.\n\nLos hogares que devuelve se releen de la base de datos, no se copian de la\nrespuesta del inicio de sesión: es así como una sesión abierta descubre que\na su usuario lo expulsaron de un hogar o lo ascendieron a administrador.\n\n**Los refresh token rotan**: el que se presenta queda revocado en la misma\noperación. Reutilizar uno ya gastado devuelve 401 y, por precaución, revoca\ntodas las sesiones vivas del usuario, porque indica que el token circula\nfuera del cliente legítimo.\n\nEndpoint público: se autentica con el propio refresh token, no con `Authorization`.\n",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Par de tokens renovado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "examples": {
                  "renovado": {
                    "description": "renovado",
                    "value": {
                      "accessToken": "eyJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJjZWxsaWVyIiwic3ViIjoiM2YxYzlhMmUtNWI3ZC00YTEwLTljODgtMmYwZTZiNGQxYTczIn0.firma",
                      "refreshToken": "Yk9sM3RQb1JmVGpXd0hxTmJHc0ttWnhEdkFlUnVMY1E",
                      "expiresIn": 900,
                      "user": {
                        "id": "3f1c9a2e-5b7d-4a10-9c88-2f0e6b4d1a73",
                        "email": "ana.rivas@gmail.com",
                        "displayName": "Ana Rivas",
                        "avatarUrl": "https://lh3.googleusercontent.com/a/ACg8ocK",
                        "locale": "es-CL",
                        "themePreference": "SYSTEM",
                        "createdAt": "2026-08-24T20:15:30Z",
                        "households": [
                          {
                            "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                            "name": "Casa Rivas",
                            "role": "ADMIN",
                            "memberCount": 3
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta el refreshToken.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "El refresh token no existe, ya se usó, se revocó o caducó.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "refreshInvalido": {
                    "description": "refreshInvalido",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "El refresh token no es válido.",
                      "instance": "/api/v1/auth/refresh"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Cerrar sesión",
        "description": "Revoca el refresh token indicado. El access token que siga vivo caduca por\nsu cuenta en cuestión de minutos.\n\nEs idempotente: cerrar sesión con un token ya inválido responde igualmente\n204, para no revelar si ese token existió alguna vez.\n\nEndpoint público: se identifica por el propio refresh token.\n",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Sesión cerrada."
          },
          "400": {
            "description": "Falta el refreshToken.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Iniciar sesión con Google",
        "description": "Valida el ID token contra el JWKS de Google (firma, emisor, audiencia,\nvigencia y correo verificado) y devuelve credenciales de Cellier.\n\nEl usuario se identifica por el claim `sub` de Google: si ya existe se le\nactualizan nombre, avatar y correo; si no, se crea la cuenta.\n\nLa respuesta incluye los hogares del usuario, de modo que el cliente puede\ndecidir a dónde llevarlo —a un hogar o a la pantalla de bienvenida— sin una\nsegunda llamada.\n\nEndpoint público: no requiere `Authorization`.\n",
        "operationId": "loginWithGoogle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "examples": {
                  "sesionIniciada": {
                    "description": "sesionIniciada",
                    "value": {
                      "accessToken": "eyJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJjZWxsaWVyIiwic3ViIjoiM2YxYzlhMmUtNWI3ZC00YTEwLTljODgtMmYwZTZiNGQxYTczIn0.firma",
                      "refreshToken": "Yk9sM3RQb1JmVGpXd0hxTmJHc0ttWnhEdkFlUnVMY1E",
                      "expiresIn": 900,
                      "user": {
                        "id": "3f1c9a2e-5b7d-4a10-9c88-2f0e6b4d1a73",
                        "email": "ana.rivas@gmail.com",
                        "displayName": "Ana Rivas",
                        "avatarUrl": "https://lh3.googleusercontent.com/a/ACg8ocK",
                        "locale": "es-CL",
                        "themePreference": "SYSTEM",
                        "createdAt": "2026-08-24T20:15:30Z",
                        "households": [
                          {
                            "id": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98",
                            "name": "Casa Rivas",
                            "role": "ADMIN",
                            "memberCount": 3
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta el idToken o el cuerpo no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "faltaIdToken": {
                    "description": "faltaIdToken",
                    "value": {
                      "type": "https://cellier.app/problems/bad-request",
                      "title": "Datos inválidos",
                      "status": 400,
                      "detail": "Uno o más campos del cuerpo de la petición no son válidos.",
                      "instance": "/api/v1/auth/google",
                      "errors": {
                        "idToken": "El idToken es obligatorio"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "El ID token de Google no es válido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "tokenInvalido": {
                    "description": "tokenInvalido",
                    "value": {
                      "type": "https://cellier.app/problems/unauthorized",
                      "title": "No autenticado",
                      "status": 401,
                      "detail": "El ID token de Google no es válido.",
                      "instance": "/api/v1/auth/google"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "RefreshRequest": {
        "type": "object",
        "description": "Refresh token vigente que se desea canjear.",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token opaco entregado en el último inicio de sesión o refresco.",
            "example": "Yk9sM3RQb1JmVGpXd0hxTmJHc0ttWnhEdkFlUnVMY1E",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "description": "Credenciales emitidas por Cellier tras autenticar al usuario.",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Access token JWT. Se envía como `Authorization: Bearer <token>`.",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJjZWxsaWVyIn0.firma"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token opaco. Es el único momento en que se entrega en claro: el servidor\nsolo guarda su hash. Consérvalo para llamar a `/api/v1/auth/refresh`.",
            "example": "Yk9sM3RQb1JmVGpXd0hxTmJHc0ttWnhEdkFlUnVMY1E"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Segundos de vigencia del access token.",
            "example": 900
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile",
            "description": "Perfil del usuario autenticado."
          }
        }
      },
      "HouseholdSummary": {
        "type": "object",
        "description": "Hogar del usuario, con su rol y el tamaño del grupo.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador del hogar.",
            "example": "8c2b7e14-9a3d-4f60-b1c5-0d7e2a6f4b98"
          },
          "name": {
            "type": "string",
            "description": "Nombre del hogar.",
            "example": "Casa Rivas"
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario autenticado en este hogar.",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64",
            "description": "Número de personas que pertenecen al hogar.",
            "example": 3
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "description": "Perfil del usuario autenticado, con los hogares a los que pertenece.\n\nEs la misma forma en `/api/v1/me`, en el inicio de sesión y en la renovación de\ntokens: el cliente recibe los hogares junto con las credenciales y puede decidir a\ndónde llevar al usuario sin una segunda llamada.\n",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador del usuario en Cellier.",
            "example": "3f1c9a2e-5b7d-4a10-9c88-2f0e6b4d1a73"
          },
          "email": {
            "type": "string",
            "description": "Correo de la cuenta de Google asociada.",
            "example": "ana.rivas@gmail.com"
          },
          "displayName": {
            "type": "string",
            "description": "Nombre para mostrar.",
            "example": "Ana Rivas"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL del avatar, o null si la cuenta no tiene foto.",
            "example": "https://lh3.googleusercontent.com/a/ACg8ocK"
          },
          "locale": {
            "type": "string",
            "description": "Configuración regional preferida.",
            "example": "es-CL"
          },
          "themePreference": {
            "type": "string",
            "description": "Preferencia de tema de la interfaz.",
            "enum": [
              "SYSTEM",
              "LIGHT",
              "DARK"
            ],
            "example": "SYSTEM"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de alta en Cellier.",
            "example": "2026-08-24T20:15:30Z"
          },
          "households": {
            "type": "array",
            "description": "Hogares a los que pertenece, con su rol en cada uno. Lista vacía si aún no\npertenece a ninguno, que es la señal de que toca la pantalla de bienvenida.\n\nSe relee de la base de datos en cada respuesta, también al renovar tokens: es\nasí como el cliente se entera de que lo expulsaron de un hogar o de que lo\nascendieron a administrador, sin que el rol tenga que viajar dentro del token.\n\nNo incluye el código de ingreso: ese dato solo sale por\n`GET /api/v1/households/{householdId}`.",
            "items": {
              "$ref": "#/components/schemas/HouseholdSummary"
            }
          }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "description": "Refresh token de la sesión que se quiere cerrar.",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token a revocar.",
            "example": "Yk9sM3RQb1JmVGpXd0hxTmJHc0ttWnhEdkFlUnVMY1E",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "GoogleLoginRequest": {
        "type": "object",
        "description": "ID token obtenido de Google Identity Services en el cliente.",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "ID token (JWT) que devuelve Google Identity Services tras el inicio de sesión.",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjFhIn0.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20ifQ.firma",
            "minLength": 1
          }
        },
        "required": [
          "idToken"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Token JWT emitido por Cellier.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}